Moved SearchRankingCompareMetric to a subfolder in `widgets`
